requirements.txt
faiss-cpu
jsonlines
rank_bm25
jieba
jsonlines
nltk
numpy
ollama
openai
rouge
rouge-chinese
sentence-transformers
tqdm
PyYAML
pydantic
pysbd
langchain
langchain-experimental
langchain-huggingface
langchain_ollama
langchain-text-splitters
